+++ tests/correlation_stats_binary_test.py
@@ -24,9 +24,9 @@ features_raw = load_breast_cancer().feature_names
 features = []
 
 for feature in features_raw:
-    features.append(Feature(feature, "float", feature))
-features.append(Feature("race", "integer", "race value", categorical=True, values={0: "black", 1: "white"}))
-features.append(Feature("gender", "integer", "race value", categorical=True, values={1: "male", 0: "female"}))
+    features.append(Feature(feature, "numeric", feature))
+features.append(Feature("race", "numeric", "race value", categorical=True, values={0: "black", 1: "white"}))
+features.append(Feature("gender", "numeric", "race value", categorical=True, values={1: "male", 0: "female"}))
 
 training_data = NumpyData(xTrain, yTrain)
 test_data = NumpyData(xTest, yTest)
@@ -34,7 +34,7 @@ dataset = Dataset({"train": training_data, "test": test_data})
 meta = MetaDatabase(features)
 
 rfc = RandomForestClassifier(n_estimators=10, criterion='entropy', random_state=0)
-output = Feature("Cancer Prediction", "integer", "Cancer Prediction", categorical=True,
+output = Feature("Cancer Prediction", "numeric", "Cancer Prediction", categorical=True,
                  values={0: "No Cancer", 1: "Cancer"})
 model = Model(agent=rfc, output_features=output, name="cisco_cancer_ai", predict_fun=rfc.predict,
               predict_prob_fun=rfc.predict_proba, model_class="Random Forest Classifier")
